--- Assets/00WorkSpace/SJH/Scripts/PokemonSkillScripts/AOESkill.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 31ea02e0f0b2e5a4cbe766d923b7ecd3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
