+++ Assets/00WorkSpace/SJH/Scripts/PokemonSkillScripts/AOE_BlastBurn.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 16686566367b930469055061e548fe8b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
