+++ Assets/00WorkSpace/SJH/Scripts/PokemonSkillScripts/BlastBurn.cs
@@ -6,7 +6,7 @@ public class BlastBurn : IAttack
 	public void Attack(Transform attacker, Vector2 attackDir, BattleDataTable attackerData, PokemonSkill skill)
 	{
 		GameObject go = PhotonNetwork.Instantiate($"PokemonSkillPrefabs/{skill.EffectPrefab.name}", attacker.transform.position, Quaternion.identity);
-		AOESkill aoe = go.GetComponent<AOESkill>();
+		AOE_BlastBurn aoe = go.GetComponent<AOE_BlastBurn>();
 		if (aoe != null) aoe.Init(attacker, attackDir, attackerData, skill);
 		else
 		{
